+++ apps/hashdotai/scripts/sync-algolia-index.ts
@@ -1,8 +1,9 @@
 import fs from "node:fs";
 import path from "node:path";
-import matter from "gray-matter";
+
 import algoliasearch from "algoliasearch";
 import * as envalid from "envalid";
+import matter from "gray-matter";
 
 const monorepoDirPath = path.resolve(__dirname, "../../..");
