+++ apps/hashdotdev/src/components/calculation-block.tsx
