+++ apps/hashdotdev/src/components/gradient-container.tsx
