+++ apps/hashdotdev/src/components/hiring-banner.tsx
@@ -1,8 +1,8 @@
 import { faArrowRight, faBullhorn } from "@fortawesome/free-solid-svg-icons";
 import { Box, Container, Typography } from "@mui/material";
 
-import { FontAwesomeIcon } from "./icons/FontAwesomeIcon";
-import { Link } from "./Link";
+import { FontAwesomeIcon } from "./icons/font-awesome-icon";
+import { Link } from "./xxx-link";
 
 const CAREERS_SITE = "https://hash.ai/careers";
