+++ apps/hashdotdev/src/components/icons/fa-icon.tsx
