+++ apps/hashdotdev/src/components/image-with-text.tsx
@@ -1,6 +1,7 @@
 import { Box } from "@mui/material";
 import { FunctionComponent, ReactNode } from "react";
-import { mdxImageClasses } from "./MdxImage";
+
+import { mdxImageClasses } from "./mdx-image";
 
 export const ImageWithText: FunctionComponent<{ children?: ReactNode }> = ({
   children,
