+++ apps/hashdotdev/src/components/mdx-image.tsx
@@ -1,8 +1,9 @@
 import { Box } from "@mui/system";
 import { ImageProps } from "next/dist/client/legacy/image";
 import Image from "next/legacy/image";
-import { HTMLProps, FunctionComponent } from "react";
-import { useBlogPostPhotos } from "./BlogPost";
+import { FunctionComponent, HTMLProps } from "react";
+
+import { useBlogPostPhotos } from "./blog-post";
 
 export const mdxImageClasses = { root: "MdxImage" };
