+++ apps/hashdotdev/src/components/mdx-page-content.tsx
@@ -1,7 +1,8 @@
 // @todo update from blockprotocol
 import { MDXRemote, MDXRemoteSerializeResult } from "next-mdx-remote";
 import { FunctionComponent } from "react";
-import { mdxComponents } from "../util/mdxComponents";
+
+import { mdxComponents } from "../util/mdx-components";
 
 type MdxPageContentProps = {
   serializedPage: MDXRemoteSerializeResult<Record<string, unknown>>;
