+++ apps/hashdotdev/src/components/xxx-snippet.tsx
