+++ apps/hashdotdev/src/pages/index.page.tsx
@@ -10,10 +10,11 @@ import { useTheme } from "@mui/system";
 import type { NextPage } from "next";
 import Image from "next/legacy/image";
 import { ComponentProps, FunctionComponent, ReactNode } from "react";
-import { Button } from "../components/Button";
-import { GradientContainer } from "../components/GradientContainer";
-import { FaIcon } from "../components/icons/FaIcon";
-import { Link } from "../components/Link";
+
+import { GradientContainer } from "../components/gradient-container";
+import { FaIcon } from "../components/icons/fa-icon";
+import { Button } from "../components/xxx-button";
+import { Link } from "../components/xxx-link";
 
 const StylishDivider: FunctionComponent<
   ComponentProps<typeof Stack> & { wide?: boolean }
