+++ apps/hashdotdev/src/theme/components/data-display/mui-avatar-theme-options.ts
