+++ apps/hashdotdev/src/theme/components/data-display/mui-icon-theme-options.ts
