+++ apps/hashdotdev/src/theme/index.ts
@@ -2,9 +2,9 @@
 
 import { createTheme } from "@mui/material";
 
+import { components } from "./components";
 import { palette } from "./palette";
 import { typography } from "./typography";
-import { components } from "./components";
 
 export const theme = createTheme({
   palette,
