+++ apps/hashdotdev/src/theme/typography.ts
@@ -1,4 +1,5 @@
 import { ThemeOptions } from "@mui/material";
+
 import { customColors } from "./palette";
 
 const fallbackFonts = [`"Helvetica"`, `"Arial"`, "sans-serif"];
