+++ apps/hashdotdev/src/util/next-types.ts
